+++ withdrawal-engine-domain/src/main/java/za/co/momentum/repo/WithdrawalRepository.java
@@ -3,5 +3,5 @@ package za.co.momentum.repo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import za.co.momentum.model.Withdrawal;
 
-public interface WithdrawalRepository extends JpaRepository<Withdrawal, Long> {
+public interface WithdrawalRepository extends JpaRepository<Withdrawal, Integer> {
 }
